+++ modules/reitit-frontend/src/reitit/frontend/history.cljs
@@ -40,7 +40,7 @@
     nil)
   (-on-navigate [this path]
     (reset! last-fragment path)
-    (on-navigate (rf/match-by-path router path) this))
+    (on-navigate (rf/match-by-path router path this) this))
   (-get-path [this]
     ;; Remove #
     ;; "" or "#" should be same as "#/"
@@ -125,7 +125,7 @@
       (-on-navigate this (-get-path this))
       this))
   (-on-navigate [this path]
-    (on-navigate (rf/match-by-path router path) this))
+    (on-navigate (rf/match-by-path router path this) this))
   (-stop [this]
     (gevents/unlistenByKey listen-key)
     (gevents/unlistenByKey click-listen-key)